use `usize` instead of `u64` in `virt_from_phys` function

+++ kernel/src/paging.rs
@@ -3,12 +3,12 @@ use lazy_static::lazy_static;
 use crate::requests::HHDM_REQUEST;
 
 lazy_static! {
-    static ref HHDM_OFFSET: u64 = HHDM_REQUEST
+    static ref HHDM_OFFSET: usize = HHDM_REQUEST
         .get_response()
         .expect("could not ask limine to get the higher half direct map offset")
-        .offset();
+        .offset() as usize;
 }
 
-pub fn virt_from_phys(phys: u64) -> u64 {
+pub fn virt_from_phys(phys: usize) -> usize {
     phys + *HHDM_OFFSET
 }
